% Generated by roxygen2 (4.0.1): do not edit by hand
\name{flatten}
\alias{flatten}
\title{Flatten an mzID related class into a table}
\usage{
flatten(object, safeNames = TRUE)
}
\arguments{
\item{object}{The object to be flattened}

\item{safeNames}{Logical. Should column names be lowered to ensure compitability between different versions of the mzIdentML schema. Defaults to TRUE}
}
\value{
A \code{data.frame} with the flattened result
}
\description{
This function flattens the content of the object into a table by merging the content intelligently (it knows the links
between the different objects).
}
\examples{
mzResults <- mzID("http://psi-pi.googlecode.com/svn/trunk/examples/1_1examples/55merge_tandem.mzid")
head(flatten(mzResults))
}
\seealso{
\code{\link{mzID-class}}
}

