% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R, R/mzIDCollection.R
\docType{methods}
\name{increment}
\alias{increment}
\alias{increment,mzIDCollection-method}
\title{Tools to handle generation of lookup names for the dictionary}
\usage{
increment(object)

\S4method{increment}{mzIDCollection}(object)
}
\arguments{
\item{object}{An mzIDCollection object}
}
\value{
An integer given the current count
}
\description{
increment looks for a counter in the data environment of mzIDCollection
objects and increments it by one if it exists. Otherwise it initialises the
counter to 1. It returns the value of the counter.
}
\section{Methods (by class)}{
\itemize{
\item \code{mzIDCollection}: Increment the counter in mzIDcollection objects
}}
\keyword{internal}

