\docType{methods}
\name{length,mzIDCollection-method}
\alias{length,mzIDCollection-method}
\title{Get the length of an mzIDCollection object}
\usage{
\S4method{length}{mzIDCollection}(x)
}
\arguments{
  \item{x}{An mzIDCollection object}
}
\value{
An integer giving the number of mzID objects contained in
the mzIDCollection
}
\description{
The length of an mzIDCollection object is defined as the
number of mzID objects it contain.
}
\seealso{
\code{\link{mzIDCollection-class}}
}

