\docType{methods}
\name{length,mzIDdatabase-method}
\alias{length,mzIDdatabase-method}
\title{Report the length of an mzIDdatabase object}
\usage{
\S4method{length}{mzIDdatabase}(x)
}
\arguments{
  \item{x}{An mzIDdatabase object}
}
\value{
A \code{numeric} giving the number of entries in the
mzIDdatabase object
}
\description{
The length of an mzIDdatabase object is defined as the
number of entries in the databse. An empty object has a
length of 0
}
\seealso{
\code{\link{mzIDdatabase-class}}
}

