\docType{methods}
\name{length,mzIDpsm-method}
\alias{length,mzIDpsm-method}
\title{Report the length of an mzIDpsm object}
\usage{
\S4method{length}{mzIDpsm}(x)
}
\arguments{
  \item{x}{An mzIDpsm object}
}
\value{
A \code{numeric} giving the number of psm's in the mzIDpsm
object
}
\description{
The length of an mzIDpsm object is defined as the number of
psm's. An empty object has a length of 0
}
\seealso{
\code{\link{mzIDpsm-class}}
}

