\name{mzID}
\alias{mzID}
\title{A class to contain data from mzIdentML-files}
\description{
  This class stores all parsed information from mzIdentML
  files
}
\details{
  The mzID class stores information in a subset of classes,
  each class having its own slot. While These classes
  should not need to be accessed directly, but descriptions
  of their content is delegated to each respective class.
}
\section{Slots}{

}

