% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mzID-getters}
\alias{database}
\alias{evidence}
\alias{files}
\alias{id}
\alias{idScanMap}
\alias{modifications}
\alias{mzID-getters}
\alias{parameters}
\alias{peptides}
\alias{scans}
\alias{software}
\title{Getter functions for identification data}
\usage{
database(object, safeNames = TRUE)

evidence(object, safeNames = TRUE)

peptides(object, safeNames = TRUE)

modifications(object)

id(object, safeNames = TRUE)

scans(object, safeNames = TRUE)

idScanMap(object)

parameters(object)

software(object)

files(object)
}
\arguments{
\item{object}{An mzID or mzIDCollection object}

\item{safeNames}{Logical. Should column names be lowered to ensure
compitability between different versions of the mzIdentML schema. Defaults to
TRUE}
}
\value{
A data frame or a list of data frames in the case of mzIDCollections
}
\description{
This set of functions are used to extract data from mzID and mzIDCollection
objects.
}
\seealso{
\code{\link{mzID-class}} \code{\link{mzIDCollection-class}}
}

