\docType{package}
\name{mzID-package}
\alias{mzID-package}
\title{A parser for the mzIdentML file format}
\description{
Using the \code{mzID} function this package is able to
parse mzIdentML files into an mzID class instance.
}
\author{
Thomas Dybdal Pedersen with contributions from Laurent
Gatto
}

