% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{removeDecoy}
\alias{removeDecoy}
\title{Remove decoy identification}
\usage{
removeDecoy(object, ...)
}
\arguments{
\item{object}{An mzID or mzIDCollection to remove decoy information from}

\item{...}{Currently ignored}
}
\value{
An mzID or mzIDCollection object depending on the input
}
\description{
This function trims down an mzID or mzIDCollection object by removing all
information that is only related to the decoy database search. If some 
information relates to both the regular and decoy database (e.g. a peptide
sequence that can be found in both databases) it is kept.
}
\seealso{
\code{\link{mzID-class}} \code{\link{mzIDCollection-class}}
}

