\docType{methods}
\name{show,mzID-method}
\alias{show,mzID-method}
\title{Show method for mzID objects}
\usage{
\S4method{show}{mzID}(object)
}
\arguments{
  \item{object}{An mzID object}
}
\value{
A description of the content of the mzID object
}
\description{
This function reports general information on the mzID
object. It is called automatically when an object is
querried.
}
\seealso{
\code{\link{mzID-class}}
}

