\docType{methods}
\name{show,mzIDCollection-method}
\alias{show,mzIDCollection-method}
\title{Show method for mzIDCollection objects}
\usage{
\S4method{show}{mzIDCollection}(object)
}
\arguments{
  \item{object}{An mzIDCollection object}
}
\description{
This function gives a very brief overview of the content of
an mzIDCollection object.
}
\seealso{
\code{\link{mzIDCollection-class}}
}

