\docType{methods}
\name{show,mzIDdatabase-method}
\alias{show,mzIDdatabase-method}
\title{Show method for mzIDdatabase objects}
\usage{
\S4method{show}{mzIDdatabase}(object)
}
\arguments{
  \item{object}{An mzIDdatabse object}
}
\value{
A description of the content of the mzIDdatabase object
}
\description{
This function reports general information on the
mzIDdatabase object. It is called automatically when an
object is querried.
}
\seealso{
\code{\link{mzIDdatabase-class}}
}

