\docType{methods}
\name{show,mzIDevidence-method}
\alias{show,mzIDevidence-method}
\title{Show method for mzIDevidence objects}
\usage{
\S4method{show}{mzIDevidence}(object)
}
\arguments{
  \item{object}{An mzIDevidence object}
}
\value{
A description of the content of the mzIDevidence object
}
\description{
This function reports general information on the
mzIDevidence object. It is called automatically when an
object is querried.
}
\seealso{
\code{\link{mzIDevidence-class}}
}

