\docType{methods}
\name{show,mzIDparameters-method}
\alias{show,mzIDparameters-method}
\title{Show method for mzIDparameters objects}
\usage{
\S4method{show}{mzIDparameters}(object)
}
\arguments{
  \item{object}{An mzIDparameters object}
}
\value{
A description of the content of the mzIDparameters object
}
\description{
This function reports general information on the
mzIDparameters object. It is called automatically when an
object is querried.
}
\seealso{
\code{\link{mzIDparameters-class}}
}

