\docType{methods}
\name{show,mzIDpeptides-method}
\alias{show,mzIDpeptides-method}
\title{Show method for mzIDpeptides objects}
\usage{
\S4method{show}{mzIDpeptides}(object)
}
\arguments{
  \item{object}{An mzIDpeptides object}
}
\value{
A description of the content of the mzIDpeptides object
}
\description{
This function reports general information on the
mzIDpeptides object. It is called automatically when an
object is querried.
}
\seealso{
\code{\link{mzIDpeptides-class}}
}

