\docType{methods}
\name{show,mzIDpsm-method}
\alias{show,mzIDpsm-method}
\title{Show method for mzIDpsm objects}
\usage{
\S4method{show}{mzIDpsm}(object)
}
\arguments{
  \item{object}{An mzIDpsm object}
}
\value{
A description of the content of the mzIDpsm object
}
\description{
This function reports general information on the mzIDpsm
object. It is called automatically when an object is
querried.
}
\seealso{
\code{\link{mzIDpsm-class}}
}

