\name{show}
\alias{show}
\title{Show method for mzIDpsm objects}
\arguments{
  \item{object}{An mzIDpsm object}

  \item{object}{An mzIDpeptides object}

  \item{object}{An mzIDparameters object}

  \item{object}{An mzIDevidence object}

  \item{object}{An mzIDdatabse object}

  \item{object}{An mzID object}
}
\value{
  A description of the content of the mzIDpsm object

  A description of the content of the mzIDpeptides object

  A description of the content of the mzIDparameters object

  A description of the content of the mzIDevidence object

  A description of the content of the mzIDdatabase object

  A description of the content of the mzID object
}
\description{
  This function reports general information on the mzIDpsm
  object. It is called automatically when an object is
  querried.

  This function reports general information on the
  mzIDpeptides object. It is called automatically when an
  object is querried.

  This function reports general information on the
  mzIDparameters object. It is called automatically when an
  object is querried.

  This function reports general information on the
  mzIDevidence object. It is called automatically when an
  object is querried.

  This function reports general information on the
  mzIDdatabase object. It is called automatically when an
  object is querried.

  This function reports general information on the mzID
  object. It is called automatically when an object is
  querried.
}
\seealso{
  \code{\link{mzIDpsm-class}}

  \code{\link{mzIDpeptides-class}}

  \code{\link{mzIDparameters-class}}

  \code{\link{mzIDevidence-class}}

  \code{\link{mzIDdatabase-class}}

  \code{\link{mzID-class}}
}

