% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{mzIDdatabase-class}
\alias{mzIDdatabase-class}
\title{A class to store database information from an mzIdentML file}
\description{
This class handles parsing and storage of database information from mzIDentML files, residing at the
/MzIdentML/SequenceCollection/DBSequence node.
}
\details{
The content of the class is stored in a data.frame with columns depending on the content of the mzIdentML
file. Required information for files conforming to the mzIdentML standard are: 'accession', 'searchDatabase_ref'
and 'id', while additional information can fx be 'length' (number of residues), 'description' (from the fasta file)
and 'sequence' (the actual sequence).
}
\section{Objects from the class}{

Objects of mzIDdatabase are not meant to be created explicitly but as part of the \code{\link{mzID-class}}. Still
object can be created with the constructor \code{\link{mzIDdatabase}} (not exported).
}

\section{Slots}{

\describe{
 \item{\code{database}:}{A data.frame containing references to all the database sequences from the mzIdentML file}
}
}

\section{Methods}{

\describe{
 \item{\code{length}:}{Reports the number of entries in the database}
}
}
\seealso{
\code{\link{mzID-class}} \code{\link{mzIDdatabase}}
}

