% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{mzIDevidence-class}
\alias{mzIDevidence-class}
\title{A class to store peptide evidence information from an mzIdentML file}
\description{
This class handles parsing and storage of peptide evidence information from mzIDentML files, residing at the
/*/x:SequenceCollection/x:PeptideEvidence node.
}
\details{
The content of the class is stored in a data.frame with columns depending on the content of the mzIdentML
file. Columns represent the attribute values of for each PeptideEvidence node. For files conforming to the HUPO
standard, dbSequence_ref, id and peptide_ref is required while start, end, pre, post, name, isDecoy, frame and
translationTable_ref are optional. Information residing in cvParam and userParam children are not parsed.
}
\section{Objects from the class}{

Objects of mzIDevidence are not meant to be created explicitly but as part of the \code{\link{mzID-class}}. Still
object can be created with the constructor \code{\link{mzIDevidence}} (not exported).
}

\section{Slots}{

\describe{
 \item{\code{evidence}:}{A data.frame containing all peptide evidence from the mzIdentML file}
}
}

\section{Methods}{

\describe{
 \item{\code{length}:}{Reports the number of peptide evidences}
}
}
\seealso{
\code{\link{mzID-class}} \code{\link{mzIDevidence}}
}

