% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{mzIDevidence-class}
\alias{evidence,mzIDevidence-method}
\alias{length,mzIDevidence-method}
\alias{mzIDevidence-class}
\alias{show,mzIDevidence-method}
\title{A class to store peptide evidence information from an mzIdentML file}
\usage{
\S4method{show}{mzIDevidence}(object)

\S4method{length}{mzIDevidence}(x)

\S4method{evidence}{mzIDevidence}(object, safeNames = TRUE)
}
\arguments{
\item{object}{An mzIDevidence object}

\item{x}{An mzIDevidence object}

\item{safeNames}{Should column names be lowercased to ensure compatibility
between v1.0 and v1.1 files?}
}
\description{
This class handles parsing and storage of peptide evidence information from
mzIDentML files, residing at the /*/x:SequenceCollection/x:PeptideEvidence
node.
}
\details{
The content of the class is stored in a data.frame with columns depending on
the content of the mzIdentML file. Columns represent the attribute values of
for each PeptideEvidence node. For files conforming to the HUPO standard,
dbSequence_ref, id and peptide_ref is required while start, end, pre, post,
name, isDecoy, frame and translationTable_ref are optional. Information
residing in cvParam and userParam children are not parsed.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Short summary of the content of the object

\item \code{length}: Report number of evidence

\item \code{evidence}: Get the evidence from the peptide search
}}
\section{Slots}{

\describe{
\item{\code{evidence}}{A data.frame containing all peptide evidence from the
mzIdentML file}
}}
\section{Objects from the class}{

Objects of mzIDevidence are not meant to be created explicitly but as part of
the \code{\link{mzID-class}}. Still object can be created with the
constructor \code{\link{mzIDevidence}}.
}
\seealso{
\code{\link{mzIDevidence}}

Other mzID.classes: \code{\link{[,mzIDCollection,character,missing,missing-method}},
  \code{\link{[,mzIDCollection,logical,missing,missing-method}},
  \code{\link{[,mzIDCollection,numeric,missing,missing-method}},
  \code{\link{[[,mzIDCollection,character,missing-method}},
  \code{\link{[[,mzIDCollection,numeric,missing-method}},
  \code{\link{as.list.mzIDCollection}},
  \code{\link{c,mzIDCollection-method}},
  \code{\link{database,mzIDCollection-method}},
  \code{\link{evidence,mzIDCollection-method}},
  \code{\link{files,mzIDCollection-method}},
  \code{\link{id,mzIDCollection-method}},
  \code{\link{idScanMap,mzIDCollection-method}},
  \code{\link{length,mzIDCollection-method}},
  \code{\link{modifications,mzIDCollection-method}},
  \code{\link{mzIDCollection-class}},
  \code{\link{names,mzIDCollection-method}},
  \code{\link{names<-,mzIDCollection,character-method}},
  \code{\link{parameters,mzIDCollection-method}},
  \code{\link{peptides,mzIDCollection-method}},
  \code{\link{removeDecoy,mzIDCollection-method}},
  \code{\link{scans,mzIDCollection-method}},
  \code{\link{show,mzIDCollection-method}},
  \code{\link{software,mzIDCollection-method}};
  \code{\link{c,mzID-method}},
  \code{\link{database,mzID-method}},
  \code{\link{evidence,mzID-method}},
  \code{\link{files,mzID-method}},
  \code{\link{id,mzID-method}},
  \code{\link{idScanMap,mzID-method}},
  \code{\link{length,mzID-method}},
  \code{\link{modifications,mzID-method}},
  \code{\link{mzID-class}},
  \code{\link{parameters,mzID-method}},
  \code{\link{peptides,mzID-method}},
  \code{\link{removeDecoy,mzID-method}},
  \code{\link{scans,mzID-method}},
  \code{\link{show,mzID-method}},
  \code{\link{software,mzID-method}};
  \code{\link{database,mzIDdatabase-method}},
  \code{\link{length,mzIDdatabase-method}},
  \code{\link{mzIDdatabase-class}},
  \code{\link{show,mzIDdatabase-method}};
  \code{\link{files,mzIDparameters-method}},
  \code{\link{length,mzIDparameters-method}},
  \code{\link{mzIDparameters-class}},
  \code{\link{parameters,mzIDparameters-method}},
  \code{\link{show,mzIDparameters-method}},
  \code{\link{software,mzIDparameters-method}};
  \code{\link{id,mzIDpsm-method}},
  \code{\link{idScanMap,mzIDpsm-method}},
  \code{\link{length,mzIDpsm-method}},
  \code{\link{mzIDpsm-class}},
  \code{\link{scans,mzIDpsm-method}},
  \code{\link{show,mzIDpsm-method}};
  \code{\link{length,mzIDpeptides-method}},
  \code{\link{modifications,mzIDpeptides-method}},
  \code{\link{mzIDpeptides-class}},
  \code{\link{peptides,mzIDpeptides-method}},
  \code{\link{show,mzIDpeptides-method}}
}

