% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mzIDpsm.R
\docType{class}
\name{mzIDpsm-class}
\alias{id,mzIDpsm-method}
\alias{idScanMap,mzIDpsm-method}
\alias{length,mzIDpsm-method}
\alias{mzIDpsm-class}
\alias{scans,mzIDpsm-method}
\alias{show,mzIDpsm-method}
\title{A class to store psm information from an mzIdentML file}
\usage{
\S4method{show}{mzIDpsm}(object)

\S4method{length}{mzIDpsm}(x)

\S4method{id}{mzIDpsm}(object, safeNames = TRUE)

\S4method{scans}{mzIDpsm}(object, safeNames = TRUE)

\S4method{idScanMap}{mzIDpsm}(object)
}
\arguments{
\item{object}{An mzIDpsm object}

\item{x}{An mzIDpsm object}

\item{safeNames}{Should column names be lowercased to ensure compatibility
between v1.0 and v1.1 files?}
}
\description{
This class handles parsing and storage of scan info and the related psm's.
This information resides in the
/*/x:DataCollection/x:AnalysisData/x:SpectrumIdentificationList/x:SpectrumIdentificationResult
node.
}
\details{
The content of the class is stored as two data frames: One containing a row
for each scan in the results, and one containing all psm's in the results.
Additionally a list containing indexing from scan to psm is stored.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: A short summary of the content

\item \code{length}: Get the number of psm'

\item \code{id}: Get the identification results

\item \code{scans}: Get the scans matched to peptides

\item \code{idScanMap}: Get the link between scans and identifications
}}
\section{Slots}{

\describe{
\item{\code{scans}}{A data.frame containing all reference to all scans with at least
one psm. The columns gives at least an ID, a spectrumID and a reference to
the file used.}

\item{\code{id}}{A data.frame containing all psm's from the analysis. The columns
depend on the file but at least id, chargeState, experimentalMassToCharge,
passThreshold and rank must exist according to the mzIdentML specifications.}

\item{\code{mapping}}{A list with an entry for each row in @scans. Each entry
contains an integer vector pointing to the related rows in @id.}
}}
\section{Objects from the class}{

Objects of mzIDpsm are not meant to be created explicitly but as part of the
\code{\link{mzID-class}}. Still object can be created with the constructor
\code{\link{mzIDpsm}}.
}
\seealso{
\code{\link{mzIDpsm}}

Other mzID.classes: \code{\link{[,mzIDCollection,character,missing,missing-method}},
  \code{\link{[,mzIDCollection,logical,missing,missing-method}},
  \code{\link{[,mzIDCollection,numeric,missing,missing-method}},
  \code{\link{[[,mzIDCollection,character,missing-method}},
  \code{\link{[[,mzIDCollection,numeric,missing-method}},
  \code{\link{as.list.mzIDCollection}},
  \code{\link{c,mzIDCollection-method}},
  \code{\link{database,mzIDCollection-method}},
  \code{\link{evidence,mzIDCollection-method}},
  \code{\link{files,mzIDCollection-method}},
  \code{\link{id,mzIDCollection-method}},
  \code{\link{idScanMap,mzIDCollection-method}},
  \code{\link{length,mzIDCollection-method}},
  \code{\link{modifications,mzIDCollection-method}},
  \code{\link{mzIDCollection-class}},
  \code{\link{names,mzIDCollection-method}},
  \code{\link{names<-,mzIDCollection,character-method}},
  \code{\link{parameters,mzIDCollection-method}},
  \code{\link{peptides,mzIDCollection-method}},
  \code{\link{removeDecoy,mzIDCollection-method}},
  \code{\link{scans,mzIDCollection-method}},
  \code{\link{show,mzIDCollection-method}},
  \code{\link{software,mzIDCollection-method}};
  \code{\link{c,mzID-method}},
  \code{\link{database,mzID-method}},
  \code{\link{evidence,mzID-method}},
  \code{\link{files,mzID-method}},
  \code{\link{id,mzID-method}},
  \code{\link{idScanMap,mzID-method}},
  \code{\link{length,mzID-method}},
  \code{\link{modifications,mzID-method}},
  \code{\link{mzID-class}},
  \code{\link{parameters,mzID-method}},
  \code{\link{peptides,mzID-method}},
  \code{\link{removeDecoy,mzID-method}},
  \code{\link{scans,mzID-method}},
  \code{\link{show,mzID-method}},
  \code{\link{software,mzID-method}};
  \code{\link{database,mzIDdatabase-method}},
  \code{\link{length,mzIDdatabase-method}},
  \code{\link{mzIDdatabase-class}},
  \code{\link{show,mzIDdatabase-method}};
  \code{\link{evidence,mzIDevidence-method}},
  \code{\link{length,mzIDevidence-method}},
  \code{\link{mzIDevidence-class}},
  \code{\link{show,mzIDevidence-method}};
  \code{\link{files,mzIDparameters-method}},
  \code{\link{length,mzIDparameters-method}},
  \code{\link{mzIDparameters-class}},
  \code{\link{parameters,mzIDparameters-method}},
  \code{\link{show,mzIDparameters-method}},
  \code{\link{software,mzIDparameters-method}};
  \code{\link{length,mzIDpeptides-method}},
  \code{\link{modifications,mzIDpeptides-method}},
  \code{\link{mzIDpeptides-class}},
  \code{\link{peptides,mzIDpeptides-method}},
  \code{\link{show,mzIDpeptides-method}}
}

